broke darray functions into internal and external files
