+++ src/clib/CMakeLists.txt
@@ -9,7 +9,7 @@ project (PIOC C)
 add_library (pioc topology.c pio_file.c pioc_support.c pio_lists.c
   pioc.c pioc_sc.c pio_spmd.c pio_rearrange.c pio_nc4.c bget.c
   pio_nc.c pio_put_nc.c pio_get_nc.c pio_msg.c pio_varm.c
-  pio_darray.c)
+  pio_darray.c pio_darray_int.c)
 
 # set up include-directories
 include_directories(
